Teslameter & LineFeed character
Investigating the output of the known MATLAB script with a logic
analyzer showed that the LF character should be "\n\0" .
It seems to be appended to the message directly, though it could be a
separate message altogether.

Assuming the hardware flow control I implemented in CppLinuxSerial does
work, this should work.