Tip Tracking - Commit 1
Added tip_tracking program that takes in a picture and applies Zhang
Suen skeletonisation algorithms to reduce it to its fundamental
component.

From there, we take the rightmost point and consider that the tip.

The main limitation here is that we consider the rightmost point the
tip.

That is fine for the current samples used for testing, but are we ever
going to be pushing the tentacle in other directions?