Tip tracking - Commit 2
Added a third screenshot to better analyze different stages of motion.

Adapted tip_tracking to work on an mp4 file (omitted to limit upload
sizes).

Moved the picture-wise implementation to skeleton_method.cpp in case any
further adjustments are needed.

Currently the skeletonisation works fine on video, but we need a more
efficient way to track the end of the tentacle (right-most, bottom-most
and combinations do not work).